+++ lib/ui/settings/SettingsSectionsContentWidget.dart
@@ -30,7 +30,6 @@ import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/service/connectivity.dart';
 import 'package:illinois/service/FirebaseMessaging.dart';
 import 'package:illinois/service/FlexUI.dart';
-import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/log.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:illinois/service/Analytics.dart';
@@ -39,32 +38,25 @@ import 'package:rokwire_plugin/service/config.dart' as rokwire;
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:illinois/ui/WebPanel.dart';
 import 'package:illinois/ui/debug/DebugHomePanel.dart';
-import 'package:illinois/ui/settings/SettingsNotificationsPanel.dart';
-import 'package:illinois/ui/settings/SettingsPersonalInformationPanel.dart';
-import 'package:illinois/ui/settings/SettingsPrivacyCenterPanel.dart';
 import 'package:illinois/ui/settings/SettingsWidgets.dart';
-import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
-import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
 import 'package:illinois/ui/widgets/RibbonButton.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:package_info/package_info.dart';
 
-class SettingsHomePanel extends StatefulWidget {
+class SettingsSectionsContentWidget extends StatefulWidget {
   @override
-  _SettingsHomePanelState createState() => _SettingsHomePanelState();
+  _SettingsSectionsContentWidgetState createState() => _SettingsSectionsContentWidgetState();
 }
 
-class _SettingsHomePanelState extends State<SettingsHomePanel> implements NotificationsListener {
+class _SettingsSectionsContentWidgetState extends State<SettingsSectionsContentWidget> implements NotificationsListener {
 
   static BorderRadius _bottomRounding = BorderRadius.only(bottomLeft: Radius.circular(5), bottomRight: Radius.circular(5));
   static BorderRadius _topRounding = BorderRadius.only(topLeft: Radius.circular(5), topRight: Radius.circular(5));
   static BorderRadius _allRounding = BorderRadius.all(Radius.circular(5));
+  static Border _allBorder = Border.all(color: Styles().colors!.surfaceAccent!, width: 1);
 
-  static BorderSide _borderSide = BorderSide(color: Styles().colors!.surfaceAccent!, width: 1);
-  static BoxBorder _topBorder = Border(top: _borderSide);
-  
   String _versionName = "";
   bool _connectingNetId = false;
 
@@ -112,21 +104,11 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   Widget build(BuildContext context) {
     
     List<Widget> contentList = [];
-    List<Widget> actionsList = [];
 
     List<dynamic> codes = FlexUI()['settings'] ?? [];
 
     for (String code in codes) {
-      if (code == 'user_info') {
-        contentList.add(_buildUserInfo());
-      }
-      else if (code == 'privacy_center') {
-        contentList.add(_buildPrivacyCenterButton(),);
-      }
-      else if (code == 'preferences') {
-        contentList.add(_buildPreferences(),);
-      }
-      else if (code == 'connect') {
+      if (code == 'connect') {
         contentList.add(_buildConnect());
       }
       else if (code == 'connected') {
@@ -138,118 +120,25 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
       else if (code == 'feedback') {
         contentList.add(_buildFeedback(),);
       }
-      else if (code == 'delete') {
-        contentList.add(_buildPrivacyDelete());
-      }
     }
 
     if (kDebugMode || (Config().configEnvironment == rokwire.ConfigEnvironment.dev)) {
       contentList.add(_buildDebug());
-      actionsList.add(_buildHeaderBarDebug());
     }
 
     contentList.add(_buildVersionInfo());
     
     contentList.add(Container(height: 12,),);
 
-    return Scaffold(
-      appBar: HeaderBar(
-        titleWidget: _buildHeaderBarTitle(),
-        actions: actionsList,
-      ),
-      body: Column(
-        children: <Widget>[
-          Expanded(
-            child: SingleChildScrollView(
-              child: Container(
-                color: Styles().colors!.background,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: contentList,
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-      backgroundColor: Styles().colors!.background,
-      bottomNavigationBar: uiuc.TabBar(),
-    );
+    return Column(crossAxisAlignment: CrossAxisAlignment.start, children: contentList);
   }
-
-  //Header Bar
-  Widget _buildHeaderBarTitle() {
-    return _DebugContainer(child:
-      Text(Localization().getStringEx("panel.settings.home.settings.header", "Settings"), style: TextStyle(color: Colors.white, fontSize: 16, fontWeight: FontWeight.w900, letterSpacing: 1.0, ), textAlign: TextAlign.center,),
-    );
-  }
-
-  //Privacy Center
-  Widget _buildPrivacyCenterButton(){
-    return GestureDetector(
-        onTap: (){
-          Analytics().logSelect(target: "Privacy Center");
-          Navigator.push(context, CupertinoPageRoute(builder: (context) =>SettingsPrivacyCenterPanel()));
-        },
-        child: Semantics(
-            button: true,
-            child:Container(
-              padding: EdgeInsets.only(left: 16, right: 16, top: 10),
-              child:Container(
-                padding: EdgeInsets.symmetric(horizontal: 12, vertical: 12),
-                decoration: BoxDecoration(
-                    color: UiColors.fromHex("9318bb"),
-                    //border: Border.all(color: Colors.grey, width: 1),
-                    borderRadius: BorderRadius.circular(6),
-                    boxShadow: [BoxShadow(color: Styles().colors!.blackTransparent018!, spreadRadius: 2.0, blurRadius: 6.0, offset: Offset(2, 2))]
-                ),
-                child:  Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: <Widget>[
-                    Expanded(child:
-                    Text(
-                      Localization().getStringEx("panel.setting.home.button.privacy_center.title","Privacy Center"), //TBD to Strings
-                      textAlign: TextAlign.start,
-                      style: TextStyle(
-                          color: Styles().colors!.white,
-                          fontSize: 20,
-                          fontFamily: Styles().fontFamilies!.bold),
-                    )),
-                    Image.asset("images/group-8.png", excludeFromSemantics: true,),
-                  ],),
-              ),
-            )));
-  }
-
-
-  // User Info
-
-  Widget _buildUserInfo() {
-    String fullName = Auth2().fullName ?? "";
-    bool hasFullName =  StringUtils.isNotEmpty(fullName);
-    String welcomeMessage = StringUtils.isNotEmpty(fullName)
-        ? AppDateTimeUtils.getDayGreeting() + ","
-        : Localization().getStringEx("panel.settings.home.user_info.title.sufix", "Welcome to Illinois");
-    return Container(
-        width: double.infinity,
-        child: Padding(
-            padding: EdgeInsets.symmetric(horizontal: 24, vertical: 12),
-            child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
-              Text(welcomeMessage, style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 20)),
-              Visibility(
-                visible: hasFullName,
-                  child: Text(fullName, style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 28))
-              ),
-            ])));
-  }
-
 
   // Connect
 
   Widget _buildConnect() {
     List<Widget> contentList =  [];
     contentList.add(Padding(
-        padding: EdgeInsets.only(left: 8, right: 8, top: 12, bottom: 2),
+        padding: EdgeInsets.only(bottom: 2),
         child: Text(
           Localization().getStringEx("panel.settings.home.connect.not_logged_in.title", "Sign in to Illinois"),
           style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 20),
@@ -261,7 +150,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     for (String code in codes) {
       if (code == 'netid') {
           contentList.add(Padding(
-            padding: EdgeInsets.all(10),
+            padding: EdgeInsets.symmetric(vertical: 10),
             child: new RichText(
               text: new TextSpan(
                 style: TextStyle(color: Styles().colors!.textBackground, fontFamily: Styles().fontFamilies!.regular, fontSize: 16),
@@ -281,7 +170,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
               ),
             )),);
           contentList.add(RibbonButton(
-            border: Border.all(color: Styles().colors!.surfaceAccent!, width: 0),
+            border: _allBorder,
             borderRadius: _allRounding,
             label: Localization().getStringEx("panel.settings.home.connect.not_logged_in.netid.title", "Sign in with your NetID"),
             progress: _connectingNetId == true,
@@ -290,7 +179,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
       }
       else if (code == 'phone_or_email') {
           contentList.add(Padding(
-            padding: EdgeInsets.all(10),
+            padding: EdgeInsets.symmetric(vertical: 10),
             child: new RichText(
               text: new TextSpan(
                 style: TextStyle(color: Styles().colors!.textBackground, fontFamily: Styles().fontFamilies!.regular, fontSize: 16),
@@ -306,14 +195,14 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
             )),);
           contentList.add(RibbonButton(
             borderRadius: _allRounding,
-            border: Border.all(color: Styles().colors!.surfaceAccent!, width: 0),
+            border: _allBorder,
             label: Localization().getStringEx("panel.settings.home.connect.not_logged_in.phone_or_email.title", "Continue"),
             onTap: _onPhoneOrEmailLoginClicked),);
       }
     }
 
     return Padding(
-      padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+      padding: EdgeInsets.symmetric(vertical: 12),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: contentList),
@@ -373,7 +262,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     return Visibility(
         visible: CollectionUtils.isNotEmpty(contentList),
         child: Padding(
-            padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+            padding: EdgeInsets.symmetric(vertical: 12),
             child: Container(
                 padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
                 decoration: BoxDecoration(
@@ -389,7 +278,6 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     List<dynamic> codes = FlexUI()['settings.connected.netid'] ?? [];
     for (int index = 0; index < codes.length; index++) {
       String code = codes[index];
-      BorderRadius borderRadius = _borderRadiusFromIndex(index, codes.length);
       if (code == 'info') {
         contentList.add(Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
               Text(Localization().getStringEx("panel.settings.home.net_id.message", "Signed in with your NetID"),
@@ -397,15 +285,6 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
               Padding(padding: EdgeInsets.only(top: 3), child: Text(Auth2().fullName ?? "",
                   style: TextStyle(color: Styles().colors!.fillColorPrimary, fontFamily: Styles().fontFamilies!.bold, fontSize: 20))),
             ]));
-      }
-      else if (code == 'authenticate') {
-        contentList.add(RibbonButton(
-          borderRadius: borderRadius,
-          border: Border.all(color: Styles().colors!.surfaceAccent!, width: 0),
-          label: Localization().getStringEx("panel.settings.home.net_id.button.connect", "Sign in with your NetID"),
-          progress: (_connectingNetId == true),
-          onTap: _onConnectNetIdClicked
-        ),);
       }
       else if (code == 'disconnect') {
         contentList.add(Padding(
@@ -432,7 +311,6 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     List<dynamic> codes = FlexUI()['settings.connected.phone'] ?? [];
     for (int index = 0; index < codes.length; index++) {
       String code = codes[index];
-      BorderRadius borderRadius = _borderRadiusFromIndex(index, codes.length);
       if (code == 'info') {
         contentList.add(Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
           Text(Localization().getStringEx("panel.settings.home.phone_ver.message", "Signed in with your Phone"),
@@ -451,8 +329,8 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
       }
       else if (code == 'verify') {
         contentList.add(RibbonButton(
-            borderRadius: borderRadius,
-            border: Border.all(color: Styles().colors!.surfaceAccent!, width: 0),
+            border: _allBorder,
+            borderRadius: _allRounding,
             label: Localization().getStringEx("panel.settings.home.phone_ver.button.connect", "Verify Your Phone Number"),
             onTap: _onPhoneOrEmailLoginClicked));
       }
@@ -480,7 +358,6 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     List<dynamic> codes = FlexUI()['settings.connected.email'] ?? [];
     for (int index = 0; index < codes.length; index++) {
       String code = codes[index];
-      BorderRadius borderRadius = _borderRadiusFromIndex(index, codes.length);
       if (code == 'info') {
         contentList.add(Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
           Text(Localization().getStringEx("panel.settings.home.email_login.message", "Signed in with your Email"),
@@ -499,8 +376,8 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
       }
       else if (code == 'login') {
         contentList.add(RibbonButton(
-            borderRadius: borderRadius,
-            border: Border.all(color: Styles().colors!.surfaceAccent!, width: 0),
+            border: _allBorder,
+            borderRadius: _allRounding,
             label: Localization().getStringEx("panel.settings.home.email_login.button.connect", "Login With Email"),
             onTap: _onPhoneOrEmailLoginClicked));
       }
@@ -734,7 +611,9 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
       String code = codes[index];
       if (code == 'netid') {
         contentList.add(RibbonButton(
-            backgroundColor: Colors.transparent,
+            backgroundColor: Styles().colors!.white,
+            border: _allBorder, 
+            borderRadius: _allRounding,
             label: Localization().getStringEx("panel.settings.home.connect.not_linked.netid.title", "Add a NetID"),
             progress: (_connectingNetId == true),
             onTap: _onLinkNetIdClicked),
@@ -742,16 +621,18 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
       }
       else if (code == 'phone') {
         contentList.add(RibbonButton(
-            border: _borderFromIndex(index, codes.length),
-            backgroundColor: Colors.transparent,
+            backgroundColor: Styles().colors!.white,
+            border: _allBorder, 
+            borderRadius: _allRounding,
             label: Localization().getStringEx("panel.settings.home.connect.not_linked.phone.title", "Add a phone number"),
             onTap: () =>
                 _onLinkPhoneOrEmailClicked(SettingsLoginPhoneOrEmailMode.phone)),);
       }
       else if (code == 'email') {
         contentList.add(RibbonButton(
-            border: _borderFromIndex(index, codes.length),
-            backgroundColor: Colors.transparent,
+            backgroundColor: Styles().colors!.white,
+            border: _allBorder, 
+            borderRadius: _allRounding,
             label: Localization().getStringEx("panel.settings.home.connect.not_linked.email.title", "Add an email address"),
             onTap: () =>
                 _onLinkPhoneOrEmailClicked(SettingsLoginPhoneOrEmailMode.email)),);
@@ -759,10 +640,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     }
 
     if (contentList.length > 0) {
-      return Container(
-          decoration: BoxDecoration(
-              border: Border.all(color: Styles().colors!.surfaceAccent!, width: 1), borderRadius: BorderRadius.all(Radius.circular(5))),
-          child: Column(children: contentList));
+      return Column(children: contentList);
     }
 
     return Container(height: 0.0,);
@@ -898,140 +776,51 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     Navigator.push(context, CupertinoPageRoute(builder: (context) => SettingsLinkedAccountPanel(linkedAccount: linked, mode: LinkAccountMode.phone,)));
   }
 
-  // Privacy
-
-  Widget _buildPreferences() {
-    List<Widget> contentList =  [];
-
-    List<dynamic> codes = FlexUI()['settings.preferences'] ?? [];
-    for (int index = 0; index < codes.length; index++) {
-      String code = codes[index];
-      if (code == 'buttons') {
-        contentList.add(_buildPreferenceButtons());
-      }
-    }
-
-    return Padding(padding: EdgeInsets.only(left: 16, right: 16), child: Column(children: contentList,));
-  }
-
-  Widget _buildPreferenceButtons() {
-
-    List<Widget> rowWidgets = <Widget>[];
-    List<Widget> colWidgets = <Widget>[];
-
-    List<dynamic> codes = FlexUI()['settings.preferences.buttons'] ?? [];
-    for (int index = 0; index < codes.length; index++) {
-      Widget? privacyButton = _buildPreferenceButton(codes[index]);
-      if (privacyButton != null) {
-          if (rowWidgets.isNotEmpty) {
-            rowWidgets.add(Container(width: 12),);
-          }
-          rowWidgets.add(Expanded(child: privacyButton));
-          
-          if (rowWidgets.length >= 3) {
-            if (colWidgets.isNotEmpty) {
-              colWidgets.add(Container(height: 12),);
-            }
-            colWidgets.add(Row(crossAxisAlignment: CrossAxisAlignment.start, children: rowWidgets));
-            rowWidgets = <Widget>[];
-          }
-      }
-    }
-
-    if (0 < rowWidgets.length) {
-      while (rowWidgets.length < 3) {
-        rowWidgets.add(Container(width: 12),);
-        rowWidgets.add(Expanded(child: Container()));
-      }
-      if (colWidgets.isNotEmpty) {
-        colWidgets.add(Container(height: 12),);
-      }
-      colWidgets.add(Row(crossAxisAlignment: CrossAxisAlignment.start, children: rowWidgets));
-    }
-
-    return Padding(padding: EdgeInsets.only(top: 12), child: Column(children: colWidgets,));
-
-    /*return Padding(padding: EdgeInsets.only(top: 12), child:
-      Row(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
-      Expanded(child: _PrivacyGridButton(
-        title: Localization().getStringEx("panel.settings.privacy_center.button.personal_information.title", "Personal Information"),
-        hint: Localization().getStringEx("panel.settings.privacy_center.button.personal_information.hint", ""),
-        iconPath: 'images/group-5.png',
-        onTap: _onTapPersonalInformation,
-      ),),
-      Container(width: 12,),
-      Expanded(child: _PrivacyGridButton(
-        title: Localization().getStringEx("panel.settings.privacy_center.button.notifications.title", "Notification Preferences"),
-        hint: Localization().getStringEx("panel.settings.privacy_center.button.notifications.", ""),
-        iconPath: 'images/notifications.png',
-        onTap: _onTapNotifications,
-      ),),
-    ]));*/
-  }
-
-  Widget? _buildPreferenceButton(String code) {
-    if (code == 'personal_information') {
-      return _PrivacyGridButton(
-        title: Localization().getStringEx("panel.settings.privacy_center.button.personal_information.title", "Personal Information"),
-        hint: Localization().getStringEx("panel.settings.privacy_center.button.personal_information.hint", ""),
-        iconPath: 'images/group-5.png',
-        onTap: _onTapPersonalInformation,
-      );
-    }
-    else if (code == 'notifications') {
-      return _PrivacyGridButton(
-        title: Localization().getStringEx("panel.settings.privacy_center.button.notifications.title", "Notification Preferences"),
-        hint: Localization().getStringEx("panel.settings.privacy_center.button.notifications.", ""),
-        iconPath: 'images/notifications.png',
-        onTap: _onTapNotifications,
-      );
-    }
-    else {
-      return null;
-    }
-  }
-
-  void _onTapPersonalInformation() {
-    Analytics().logSelect(target: "Personal Information");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => SettingsPersonalInformationPanel()));
-  }
-
-  void _onTapNotifications() {
-    Analytics().logSelect(target: "Notifications");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => SettingsNotificationsPanel()));
-  }
-
   // Feedback
 
-  Widget _buildFeedback(){
+  Widget _buildFeedback() {
+    final String rkwPlatformMacro = '{{rkw_platform_label}}';
+    final String smartHealthyInitiativeMacro = '{{smart_healthy_initiative_label}}';
+    String appDescriptionContent = Localization().getStringEx('panel.settings.home.feedback.app.description.format',
+        'The Illinois App is the official campus app for the University of Illinois Urbana-Champaign. The app is powered by the $rkwPlatformMacro. The Rowkire project is an effort of the $smartHealthyInitiativeMacro, which is funded by the Office of the Provost at the University of Illinois.');
+    int rkwPlatformMacroPosition = appDescriptionContent.indexOf(rkwPlatformMacro);
+    int smartHealthyInitiativeMacroPosition = appDescriptionContent.indexOf(smartHealthyInitiativeMacro);
+    String feedbackMsgStart = appDescriptionContent.substring(0, rkwPlatformMacroPosition);
+    String feedbackMsgMiddle = appDescriptionContent.substring((rkwPlatformMacroPosition + rkwPlatformMacro.length), smartHealthyInitiativeMacroPosition);
+    String feedbackMsgEnd = appDescriptionContent.substring(smartHealthyInitiativeMacroPosition + smartHealthyInitiativeMacro.length);
+
     return Column(
       children: <Widget>[
         Padding(
-            padding: EdgeInsets.symmetric(horizontal: 24, vertical: 12),
+            padding: EdgeInsets.symmetric(vertical: 12),
             child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
               Text(
-                Localization().getStringEx("panel.settings.home.feedback.title", "We need your ideas!"),
+                Localization().getStringEx("panel.settings.home.feedback.title", "App Feedback"),
                 style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 20),
               ),
               Container(height: 5,),
               Text(
-                Localization().getStringEx("panel.settings.home.feedback.description", "Enjoying the app? Missing something? Tap on the bottom to submit your idea."),
+                Localization().getStringEx("panel.settings.home.feedback.description", "Enjoying the app? Missing something? The University of Illinois Smart, Healthy Communities Initiative needs your ideas and input. Thank you!"),
                 style: TextStyle(fontFamily: Styles().fontFamilies!.regular,color: Styles().colors!.textBackground, fontSize: 16),
               ),
             ])
         ),
         Padding(
-          padding: EdgeInsets.symmetric(horizontal: 18, vertical: 12),
-          child: RoundedButton(
+          padding: EdgeInsets.only(bottom: 20),
+          child: RibbonButton(
+            border: _allBorder,
+            borderRadius: _allRounding,
             label: Localization().getStringEx("panel.settings.home.button.feedback.title", "Submit Feedback"),
-            hint: Localization().getStringEx("panel.settings.home.button.feedback.hint", ""),
-            backgroundColor: Styles().colors!.background,
-            fontSize: 16.0,
-            textColor: Styles().colors!.fillColorPrimary,
-            borderColor: Styles().colors!.fillColorSecondary,
-            onTap: _onFeedbackClicked,
-          ),
+            onTap: _onFeedbackClicked
+          )
         ),
+    RichText(text: TextSpan(style: TextStyle(fontFamily: Styles().fontFamilies!.regularIt, color: Styles().colors!.textBackground, fontSize: 16), children: [
+      TextSpan(text: feedbackMsgStart),
+      TextSpan(text: Localization().getStringEx('panel.settings.home.feedback.app.description.rokwire_platform.label', 'Rokwire opensource software platform'), style: TextStyle(decoration: TextDecoration.underline)),
+      TextSpan(text: feedbackMsgMiddle),
+      TextSpan(text: Localization().getStringEx('panel.settings.home.feedback.app.description.smart_healthy_initiative.label', 'Smart, Healthy Communities Initiative'), style: TextStyle(decoration: TextDecoration.underline)),
+      TextSpan(text: feedbackMsgEnd)
+    ]))
       ],
     );
   }
@@ -1072,86 +861,16 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     }
   }
 
-  // Delete Information
-
-  Widget _buildPrivacyDelete() {
-    return Padding(padding: EdgeInsets.only(left: 18, right: 18, top: 24, bottom: 12), child:
-    Column(children: <Widget>[
-      RoundedButton(
-        backgroundColor: Styles().colors!.white,
-        borderColor: Styles().colors!.white,
-        textColor: UiColors.fromHex("#f54400"),
-        fontSize: 16,
-        fontFamily: Styles().fontFamilies!.regular,
-        label: Localization().getStringEx("panel.settings.privacy_center.button.delete_data.title", "Delete My Account"),
-        hint: Localization().getStringEx("panel.settings.privacy_center.label.delete.description", "This will delete all of your personal information that was shared and stored within the app."),
-        borderShadow: [BoxShadow(color: Color.fromRGBO(19, 41, 75, 0.3), spreadRadius: 2.0, blurRadius: 8.0, offset: Offset(0, 2))],
-        onTap: _onTapDeleteData,
-      ),
-    ],),);
-  }
-
-  void _onTapDeleteData() async{
-    final String groupsSwitchTitle = "Please delete all my contributions.";
-    int userPostCount = await Groups().getUserPostCount();
-    bool contributeInGroups = userPostCount > 0;
-
-    SettingsDialog.show(context,
-        title: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.title", "Delete your account?"),
-        message: [
-          TextSpan(text: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description1", "This will ")),
-          TextSpan(text: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description2", "Permanently "),style: TextStyle(fontFamily: Styles().fontFamilies!.bold)),
-          TextSpan(text: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description3", "delete all of your information. You will not be able to retrieve your data after you have deleted it. Are you sure you want to continue?")),
-          //TBD localization
-          TextSpan(text: contributeInGroups?
-          Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description.groups", " You have contributed to Groups. Do you wish to delete all of those entries (posts, replies, and events) or leave them for others to see.") :
-          ""
-          ),
-        ],
-        options:contributeInGroups ? [groupsSwitchTitle] : null,
-        initialOptionsSelection:contributeInGroups ?  [groupsSwitchTitle] : [],
-        continueTitle: Localization().getStringEx("panel.settings.privacy_center.button.forget_info.title","Forget My Information"),
-        onContinue: (List<String> selectedValues, OnContinueProgressController progressController ){
-          progressController(loading: true);
-          if(selectedValues.contains(groupsSwitchTitle)){
-            Groups().deleteUserData();
-          }
-          _deleteUserData().then((_){
-            progressController(loading: false);
-            Navigator.pop(context);
-          });
-
-        },
-        longButtonTitle: true
-    );
-  }
-
-  Future<void> _deleteUserData() async{
-    Analytics().logAlert(text: "Remove My Information", selection: "Yes");
-    await Auth2().deleteUser();
-  }
-
   // Debug
 
   Widget _buildDebug() {
     return Padding(
-      padding: EdgeInsets.symmetric(horizontal: 18, vertical: 24),
-      child: RoundedButton(
-        label: Localization().getStringEx("panel.profile_info.button.debug.title", "Debug"),
-        hint: Localization().getStringEx("panel.profile_info.button.debug.hint", ""),
-        backgroundColor: Styles().colors!.background,
-        fontSize: 16.0,
-        textColor: Styles().colors!.fillColorPrimary,
-        borderColor: Styles().colors!.fillColorSecondary,
-        onTap: _onDebugClicked,
-      ),
-    );
-  }
-
-  Widget _buildHeaderBarDebug() {
-    return Semantics(label: Localization().getStringEx('headerbar.debug.title', 'Debug'), hint: Localization().getStringEx('headerbar.debug.hint', ''), button: true, excludeSemantics: true, child:
-      IconButton(icon: Image.asset('images/debug-white.png'), onPressed: _onDebugClicked),
-    );
+        padding: EdgeInsets.symmetric(vertical: 24),
+        child: RibbonButton(
+            border: _allBorder,
+            borderRadius: _allRounding,
+            label: Localization().getStringEx("panel.profile_info.button.debug.title", "Debug"),
+            onTap: _onDebugClicked));
   }
 
   void _onDebugClicked() {
@@ -1159,14 +878,13 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     Navigator.push(context, CupertinoPageRoute(builder: (context) => DebugHomePanel()));
   }
 
-  //Version Info
-  Widget _buildVersionInfo(){
+  // Version Info
+  Widget _buildVersionInfo() {
+    String versionLabel = Localization().getStringEx('panel.settings.home.version.info.label', 'Illinois App Version:') + _versionName;
     return Container(
-      alignment: Alignment.center,
-      child:  Text(
-        "Version: $_versionName",
-        style: TextStyle(color: Styles().colors!.textBackground, fontFamily: Styles().fontFamilies!.regular, fontSize: 16),
-    ));
+        alignment: Alignment.center,
+        child: Text(versionLabel,
+            style: TextStyle(color: Styles().colors!.textBackground, fontFamily: Styles().fontFamilies!.regular, fontSize: 16)));
   }
 
   void _loadVersionInfo() async {
@@ -1193,15 +911,6 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     }
     else {
       return BorderRadius.zero;
-    }
-  }
-
-  BoxBorder? _borderFromIndex(int index, int length) {
-    int first = 0;
-    if (index > first) {
-      return _topBorder;
-    } else {
-      return null;
     }
   }
 
@@ -1232,7 +941,7 @@ class _OptionsSection extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-        padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+        padding: EdgeInsets.symmetric(vertical: 12),
         child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
           Visibility(
             visible: StringUtils.isNotEmpty(title),
@@ -1262,40 +971,6 @@ class _OptionsSection extends StatelessWidget {
             )
           ])
         ]));
-  }
-}
-
-class _PrivacyGridButton extends StatelessWidget {
-  final String? title;
-  final String? hint;
-  final String? iconPath;
-  final void Function()? onTap;
-
-  const _PrivacyGridButton({Key? key, this.title, this.hint, this.iconPath, this.onTap}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(onTap: onTap, child:
-      Semantics(label: title, hint: hint, button: true, excludeSemantics: true, child:
-        Padding(padding: EdgeInsets.all(2), child:
-          Container(
-            decoration: BoxDecoration(color: (Styles().colors!.white),
-              borderRadius: BorderRadius.circular(4),
-              border: Border.all(color: Colors.white, width: 2),
-              boxShadow: [BoxShadow(color: Color.fromRGBO(19, 41, 75, 0.3), spreadRadius: 2.0, blurRadius: 8.0, offset: Offset(0, 2))]),
-            child: Padding(padding: EdgeInsets.only(top: 16, bottom: 16), child:
-              Column(mainAxisAlignment: MainAxisAlignment.center, mainAxisSize: MainAxisSize.min, children: <Widget>[
-                Padding(padding: EdgeInsets.only(bottom: 16), child:
-                  Image.asset(iconPath!),
-                ),
-                Padding(padding: EdgeInsets.only(left: 10, right: 10, top: 10), child:
-                  Text(title!, textAlign: TextAlign.center, style: TextStyle(fontFamily: Styles().fontFamilies!.bold, fontSize: 16, color: Styles().colors!.fillColorPrimary)))
-              ],),
-            ),
-          ),
-        ),
-      ),
-    );
   }
 }